requirements.txt
scapy
discord.py
flask
netifaces
python-nmap
colorama
python-dotenv